requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
pydantic
pytest
httpx
psycopg2-binary
